It Works!!
It draws stuff and seems to decode correctly. Also added pinch zoom and drag to the canvas for legibility. I think the bluetooth connection Service is incomplete, I probably should update state changes to reflect bonded, etc. Need to add cursors and figure out blinking things and clean out a bunch of commented old code. We don't use some of the functionality described in the manual, so I'm not sure how to test it. Also need to get an alternative for LocalBroadcastManager